+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.3
+- Added support for latest version of Pulsar 0.4.5.
+- Added new CDN template with tailwind support.
+
 ## 0.2.2
 - Added support for latest version of Pulsar 0.4.4+1